lanchain_small/requirements.txt
langchain
langchain-community
openai
python-dotenv
faiss-cpu
jq
tiktoken
langchain-openai